building data model based on class demo toucable list, about halfway done
