--- app/src/main/java/com/android/sooz/filterapp/SpeciesData.java
@@ -1,229 +1,0 @@
-package com.android.sooz.filterapp;
-
-import java.util.ArrayList;
-
-public class SpeciesData {
-    ArrayList<String> species = new ArrayList<>();
-    species = 		{
-        "name": "Hutt",
-                "classification": "gastropod",
-                "designation": "sentient",
-                "average_height": "300",
-                "skin_colors": "green, brown, tan",
-                "hair_colors": "n/a",
-                "eye_colors": "yellow, red",
-                "average_lifespan": "1000",
-                "homeworld": "https://swapi.co/api/planets/24/",
-                "language": "Huttese",
-                "people": [
-        "https://swapi.co/api/people/16/"
-			],
-        "films": [
-        "https://swapi.co/api/films/3/",
-                "https://swapi.co/api/films/1/"
-			],
-        "created": "2014-12-10T17:12:50.410000Z",
-                "edited": "2014-12-20T21:36:42.146000Z",
-                "url": "https://swapi.co/api/species/5/"
-    },
-    {
-        "name": "Yoda's species",
-            "classification": "mammal",
-            "designation": "sentient",
-            "average_height": "66",
-            "skin_colors": "green, yellow",
-            "hair_colors": "brown, white",
-            "eye_colors": "brown, green, yellow",
-            "average_lifespan": "900",
-            "homeworld": "https://swapi.co/api/planets/28/",
-            "language": "Galactic basic",
-            "people": [
-        "https://swapi.co/api/people/20/"
-			],
-        "films": [
-        "https://swapi.co/api/films/2/",
-                "https://swapi.co/api/films/5/",
-                "https://swapi.co/api/films/4/",
-                "https://swapi.co/api/films/6/",
-                "https://swapi.co/api/films/3/"
-			],
-        "created": "2014-12-15T12:27:22.877000Z",
-            "edited": "2014-12-20T21:36:42.148000Z",
-            "url": "https://swapi.co/api/species/6/"
-    },
-    {
-        "name": "Trandoshan",
-            "classification": "reptile",
-            "designation": "sentient",
-            "average_height": "200",
-            "skin_colors": "brown, green",
-            "hair_colors": "none",
-            "eye_colors": "yellow, orange",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/29/",
-            "language": "Dosh",
-            "people": [
-        "https://swapi.co/api/people/24/"
-			],
-        "films": [
-        "https://swapi.co/api/films/2/"
-			],
-        "created": "2014-12-15T13:07:47.704000Z",
-            "edited": "2014-12-20T21:36:42.151000Z",
-            "url": "https://swapi.co/api/species/7/"
-    },
-    {
-        "name": "Mon Calamari",
-            "classification": "amphibian",
-            "designation": "sentient",
-            "average_height": "160",
-            "skin_colors": "red, blue, brown, magenta",
-            "hair_colors": "none",
-            "eye_colors": "yellow",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/31/",
-            "language": "Mon Calamarian",
-            "people": [
-        "https://swapi.co/api/people/27/"
-			],
-        "films": [
-        "https://swapi.co/api/films/3/"
-			],
-        "created": "2014-12-18T11:09:52.263000Z",
-            "edited": "2014-12-20T21:36:42.153000Z",
-            "url": "https://swapi.co/api/species/8/"
-    },
-    {
-        "name": "Ewok",
-            "classification": "mammal",
-            "designation": "sentient",
-            "average_height": "100",
-            "skin_colors": "brown",
-            "hair_colors": "white, brown, black",
-            "eye_colors": "orange, brown",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/7/",
-            "language": "Ewokese",
-            "people": [
-        "https://swapi.co/api/people/30/"
-			],
-        "films": [
-        "https://swapi.co/api/films/3/"
-			],
-        "created": "2014-12-18T11:22:00.285000Z",
-            "edited": "2014-12-20T21:36:42.155000Z",
-            "url": "https://swapi.co/api/species/9/"
-    },
-    {
-        "name": "Sullustan",
-            "classification": "mammal",
-            "designation": "sentient",
-            "average_height": "180",
-            "skin_colors": "pale",
-            "hair_colors": "none",
-            "eye_colors": "black",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/33/",
-            "language": "Sullutese",
-            "people": [
-        "https://swapi.co/api/people/31/"
-			],
-        "films": [
-        "https://swapi.co/api/films/3/"
-			],
-        "created": "2014-12-18T11:26:20.103000Z",
-            "edited": "2014-12-20T21:36:42.157000Z",
-            "url": "https://swapi.co/api/species/10/"
-    },
-    {
-        "name": "Neimodian",
-            "classification": "unknown",
-            "designation": "sentient",
-            "average_height": "180",
-            "skin_colors": "grey, green",
-            "hair_colors": "none",
-            "eye_colors": "red, pink",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/18/",
-            "language": "Neimoidia",
-            "people": [
-        "https://swapi.co/api/people/33/"
-			],
-        "films": [
-        "https://swapi.co/api/films/4/"
-			],
-        "created": "2014-12-19T17:07:31.319000Z",
-            "edited": "2014-12-20T21:36:42.160000Z",
-            "url": "https://swapi.co/api/species/11/"
-    },
-    {
-        "name": "Gungan",
-            "classification": "amphibian",
-            "designation": "sentient",
-            "average_height": "190",
-            "skin_colors": "brown, green",
-            "hair_colors": "none",
-            "eye_colors": "orange",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/8/",
-            "language": "Gungan basic",
-            "people": [
-        "https://swapi.co/api/people/36/",
-                "https://swapi.co/api/people/37/",
-                "https://swapi.co/api/people/38/"
-			],
-        "films": [
-        "https://swapi.co/api/films/5/",
-                "https://swapi.co/api/films/4/"
-			],
-        "created": "2014-12-19T17:30:37.341000Z",
-            "edited": "2014-12-20T21:36:42.163000Z",
-            "url": "https://swapi.co/api/species/12/"
-    },
-    {
-        "name": "Toydarian",
-            "classification": "mammal",
-            "designation": "sentient",
-            "average_height": "120",
-            "skin_colors": "blue, green, grey",
-            "hair_colors": "none",
-            "eye_colors": "yellow",
-            "average_lifespan": "91",
-            "homeworld": "https://swapi.co/api/planets/34/",
-            "language": "Toydarian",
-            "people": [
-        "https://swapi.co/api/people/40/"
-			],
-        "films": [
-        "https://swapi.co/api/films/5/",
-                "https://swapi.co/api/films/4/"
-			],
-        "created": "2014-12-19T17:48:56.893000Z",
-            "edited": "2014-12-20T21:36:42.165000Z",
-            "url": "https://swapi.co/api/species/13/"
-    },
-    {
-        "name": "Dug",
-            "classification": "mammal",
-            "designation": "sentient",
-            "average_height": "100",
-            "skin_colors": "brown, purple, grey, red",
-            "hair_colors": "none",
-            "eye_colors": "yellow, blue",
-            "average_lifespan": "unknown",
-            "homeworld": "https://swapi.co/api/planets/35/",
-            "language": "Dugese",
-            "people": [
-        "https://swapi.co/api/people/41/"
-			],
-        "films": [
-        "https://swapi.co/api/films/4/"
-			],
-        "created": "2014-12-19T17:53:11.214000Z",
-            "edited": "2014-12-20T21:36:42.167000Z",
-            "url": "https://swapi.co/api/species/14/"
-    }
-	]
-}
-
-}
